refac: вынести в переменные размеры экранов, длительность короткометражек, кол-во карточек.
