+++ src/constants/desktopMinWidth.js
@@ -1,0 +1,3 @@
+const desktopMinWidth = 1280;
+
+export default desktopMinWidth;
