+++ src/constants/extraCardsNumber.js
@@ -1,0 +1,3 @@
+const extraCardsNumber = { desktop: 4, tablet: 2, phone: 2 };
+
+export default extraCardsNumber;
