+++ src/constants/initialCardsAmount.js
@@ -1,0 +1,3 @@
+const initialCardsAmount = { desktop: 16, tablet: 8, phone: 5 };
+
+export default initialCardsAmount;
