+++ src/constants/shortMovieMaxDuration.js
@@ -1,0 +1,3 @@
+const shortMovieMaxDuration = 40;
+
+export default shortMovieMaxDuration;
