+++ src/constants/tabletMinWidth.js
@@ -1,0 +1,3 @@
+const tabletMinWidth = 768;
+
+export default tabletMinWidth;
